Add support for log agent.
